Fix: Race-condition in the opening of the RSA keyfile
Previously, the file was opened, and then the permissions were set, opening a
window for a race-condition in which an attacker could open the file in between
the call to `open()`, and the `os.chmod()` call.

The fix is to create the file in one atomic action, using `os.open` as the
argument to `open()`, with the `os.O_EXCL` flag, to hinder any unwanted reads of
the file.

Changelog: Title
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit e76867eb9fdc1152cf5ee6e209b1d440843683e2)